company-device-serivce/requirements.txt
anyio==3.3.4
asgi-lifespan==1.0.1
asgiref==3.5.0
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==36.0.1
dnspython==2.2.0
ecdsa==0.17.0
fastapi==0.71.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.20.0
idna==3.3
iniconfig==1.1.1
mongoengine==0.23.1
packaging==21.2
passlib==1.7.4
pluggy==1.0.0
py==1.10.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pymongo==3.12.3
pyparsing==2.4.7
pytest==6.2.5
pytest-asyncio==0.16.0
python-jose==3.3.0
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing-extensions==4.0.1
uvicorn==0.17.0.post1
